feat(stabilization): Complete Phase A-E - Database, Navigation, Features, i18n, Tests
MAJOR STABILIZATION COMPLETE:

Database (Phase A1):
- Migration v50→51 with comprehensive schema normalization
- Fixed @ColumnInfo(defaultValue) mismatches in 31 entity files
- Added missing index_scanned_receipts_matchStatus to ScannedReceipt
- All 33 tables now have consistent schema

Navigation (Phase B):
- Unified NavigationController with navigateToTab() for all 6 tabs
- Synchronized selectedTab with NavigationController.currentDestination
- Bottom bar hides on feature screens
- All 22 feature destinations properly wired

Features Menu (Phase C):
- Config-driven FeaturesMenu using FeatureConfig.allFeatures
- All 22 features visible with NEW badges for Currency, Subscription,
  Tax, Export, Manual Recurring, and Shared Groups
- Single onNavigateToFeature callback replacing 16 individual callbacks

i18n String Externalization (Phase D + G07-G08):
- Added 85+ new string resources to strings.xml
- Externalized ~330 strings across 8 screens:
  * DebugScreen (G07) - 110 strings
  * PriceProtection, SavingsGoals, WarrantyTracker (G08)
  * ReceiptMatching, SplitTemplates, CategoryScreen, CashFlowCalendar
- Fixed format string crashes (ComputeDashboardWidgetsUseCase)
- Externalized HomeViewModel strings (G06 completion)

Testing (Phase E):
- Fixed 7 unit test failures:
  * HomeViewModelStressTest - Application parameter + type fixes
  * TransactionRollbackTest - ThreadLocal transaction tracking
  * GroupTransactionCoordinatorTest - Foreign key constraint fixes
- All tests now pass (BUILD SUCCESSFUL)

New Features:
- CSV Import utility (CsvExpenseImporter.kt) for migrating old data
- Import from CSV button in Debug Menu
- Database export/import functionality in Debug Menu

Build Status:
- Kotlin compilation: SUCCESS
- Unit tests: PASSING
- Resources merge: SUCCESS

Breaking Changes: None (backward compatible)

Files Modified: 71 files
New Files: 4 (CsvExpenseImporter.kt, QA_TEST_MATRIX.md,
              STABILIZATION_SUMMARY.md, session-ses_2bbe.md)

+++ app/src/main/java/com/yourname/expensetracker/domain/usecase/dashboard/ComputeDashboardWidgetsUseCase.kt
@@ -495,11 +495,11 @@ class ComputeDashboardWidgetsUseCase @Inject constructor(
             val diff = monthSpent - previousMonthTotal
             return when {
                 diff < 0 -> Pair(
-                    UiText.from(R.string.widget_insight_spent_less_format, -diff.toInt()),
+                    UiText.from(R.string.widget_insight_spent_less_format, -diff),
                     "📉"
                 )
                 diff > previousMonthTotal * 0.2 -> Pair(
-                    UiText.from(R.string.widget_insight_spent_higher_format, diff.toInt()),
+                    UiText.from(R.string.widget_insight_spent_higher_format, diff),
                     "📈"
                 )
                 else -> null
